% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R, R/temp_conversion.R
\name{F_to_C}
\alias{F_to_C}
\title{Scatterplot}
\usage{
F_to_C(F_temp)

F_to_C(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}

\item{miRNA_name}{The miRNA}

\item{mRNA}{the mRNA}

\item{cut_off}{set at 0 and is the minimum expression it should have}
}
\value{
A scatterplot with the tissueType on the right

The temperature in degrees Celsius
}
\description{
Plot a chosen miRNA and mRNA

Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
