% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_regular.R
\docType{data}
\name{gene_expression_regular}
\alias{gene_expression_regular}
\title{gene_expression_regular}
\format{
An object of class \code{data.frame} with 492593 rows and 65 columns.
}
\usage{
gene_expression_regular
}
\description{
gene_expression_regular
}
\keyword{datasets}
