% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression_xi_hat.R
\docType{data}
\name{gene_expression_xi_hat}
\alias{gene_expression_xi_hat}
\title{gene_expression_xi_hat}
\format{
An object of class \code{data.frame} with 425389 rows and 65 columns.
}
\usage{
gene_expression_xi_hat
}
\description{
gene_expression_xi_hat
}
\keyword{datasets}
