% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNA_expression_regular.R
\docType{data}
\name{microRNA_expression_regular}
\alias{microRNA_expression_regular}
\title{microRNA_expression_regular}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 91640 rows and 32 columns.
}
\usage{
microRNA_expression_regular
}
\description{
microRNA_expression_regular
}
\keyword{datasets}
