% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNA_expression_xi_hat.R
\docType{data}
\name{microRNA_expression_xi_hat}
\alias{microRNA_expression_xi_hat}
\title{microRNA_expression_xi_hat}
\format{
An object of class \code{data.frame} with 91640 rows and 33 columns.
}
\usage{
microRNA_expression_xi_hat
}
\description{
microRNA_expression_xi_hat
}
\keyword{datasets}
